pyproject.toml
[project]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = [
    {name = "XeindsTwo",email = "yourihelo1971@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "psycopg2 (>=2.9.12,<3.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "django-image-uploader-widget (>=1.1.1,<2.0.0)",
    "faker (>=40.15.0,<41.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
